requirements.txt
Appium-Python-Client==0.9
enum34==1.0
py==1.4.23
pytest==2.6.1
selenium==2.42.1
wsgiref==0.1.2
